+++ client/src/main.ts
@@ -1,3 +1,12 @@
-import { createGame } from "./game.js";
+import Phaser from "phaser";
+import { WORLD_WIDTH, WORLD_HEIGHT } from "shared";
+import { LobbyScene } from "./lobby.js";
+import { GameScene } from "./game.js";
 
-createGame();
+new Phaser.Game({
+  type: Phaser.AUTO,
+  width: WORLD_WIDTH,
+  height: WORLD_HEIGHT,
+  scene: [LobbyScene, GameScene],
+  parent: document.body,
+});